# openapi-specs/account.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/handles": {
      "post": {
        "operationId": "HandlesControllerV1_createHandle",
        "summary": "Request to create a new handle for an account",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandleRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Handle creation request enqueued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/handles"
        ]
      }
    },
    "/v1/handles/change": {
      "post": {
        "operationId": "HandlesControllerV1_changeHandle",
        "summary": "Request to change a handle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandleRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Handle change request enqueued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/handles"
        ]
      }
    },
    "/v1/handles/change/{newHandle}": {
      "get": {
        "operationId": "HandlesControllerV1_getChangeHandlePayload",
        "summary": "Get a properly encoded ClaimHandlePayload that can be signed.",
        "parameters": [
          {
            "name": "newHandle",
            "required": true,
            "in": "path",
            "description": "newHandle in the request",
            "example": "handle",
            "schema": {
              "minLength": 3,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned an encoded ClaimHandlePayload for signing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeHandlePayloadRequest"
                }
              }
            }
          }
        },
        "tags": [
          "v1/handles"
        ]
      }
    },
    "/v1/handles/{msaId}": {
      "get": {
        "operationId": "HandlesControllerV1_getHandle",
        "summary": "Fetch a handle given an MSA Id",
        "parameters": [
          {
            "name": "msaId",
            "required": true,
            "in": "path",
            "description": "Msa Id of requested account",
            "example": "2",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found a handle",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandleResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "v1/handles"
        ]
      }
    },
    "/v1/keys/add": {
      "post": {
        "operationId": "KeysControllerV1_addKey",
        "summary": "Add new control keys for an MSA Id",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeysRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Found public keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/keys"
        ]
      }
    },
    "/v1/keys/{msaId}": {
      "get": {
        "operationId": "KeysControllerV1_getKeys",
        "summary": "Fetch public keys given an MSA Id",
        "parameters": [
          {
            "name": "msaId",
            "required": true,
            "in": "path",
            "description": "Msa Id of requested account",
            "example": "2",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Found public keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeysResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/keys"
        ]
      }
    },
    "/v1/keys/publicKeyAgreements/getAddKeyPayload": {
      "get": {
        "operationId": "KeysControllerV1_getPublicKeyAgreementsKeyPayload",
        "summary": "Get a properly encoded StatefulStorageItemizedSignaturePayloadV2 that can be signed.",
        "parameters": [
          {
            "name": "msaId",
            "required": true,
            "in": "query",
            "description": "MSA Id representing the target of this request",
            "example": "3",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "newKey",
            "required": true,
            "in": "query",
            "description": "New public key to be added to the account (32-byte value in hex format)",
            "example": "0x0ed2f8c714efcac51ca2325cfe95637e5e0b898ae397aa365978b7348a717d0b",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned an encoded StatefulStorageItemizedSignaturePayloadV2 for signing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddNewPublicKeyAgreementPayloadRequest"
                }
              }
            }
          }
        },
        "tags": [
          "v1/keys"
        ]
      }
    },
    "/v1/keys/publicKeyAgreements": {
      "post": {
        "operationId": "KeysControllerV1_AddNewPublicKeyAgreements",
        "summary": "Request to add a new public Key",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNewPublicKeyAgreementRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Add new key request enqueued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          }
        },
        "tags": [
          "v1/keys"
        ]
      }
    },
...
  },
  "info": {
    "title": "Account Service",
    "description": "Account Service API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "HandleResponseDto": {
        "type": "object",
        "properties": {
          "base_handle": {
            "type": "string"
          },
          "canonical_base": {
            "type": "string"
          },
          "suffix": {
            "type": "number"
          }
        },
        "required": [
          "base_handle",
          "canonical_base",
          "suffix"
        ]
      },
...
      "TransactionResponse": {
        "type": "object",
        "properties": {
          "referenceId": {
            "type": "string"
          }
        },
        "required": [
          "referenceId"
        ]
      },
...
      "HandlePayloadDto": {
        "type": "object",
        "properties": {
          "baseHandle": {
            "type": "string",
            "description": "base handle in the request",
            "example": "handle",
            "minLength": 3
          },
          "expiration": {
            "type": "number",
            "description": "expiration block number for this payload",
            "example": 1
          }
        },
        "required": [
          "baseHandle",
          "expiration"
        ]
      },
      "HandleRequestDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "description": "AccountId in hex or SS58 format",
            "example": "1LSLqpLWXo7A7xuiRdu6AQPnBPNJHoQSu8DBsUYJgsNEJ4N"
          },
          "payload": {
            "$ref": "#/components/schemas/HandlePayloadDto"
          },
          "proof": {
            "type": "string",
            "description": "proof is the signature for the payload",
            "example": "0x065d733ca151c9e65b78f2ba77348224d31647e6913c44ad2765c6e8ba06f834dc21d8182447d01c30f84a41d90a8f2e58001d825c6f0d61b0afe89f984eec85"
          }
        },
        "required": [
          "accountId",
          "payload",
          "proof"
        ]
      },
      "ChangeHandlePayloadRequest": {
        "type": "object",
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/HandlePayloadDto"
          },
          "encodedPayload": {
            "type": "string",
            "description": "Raw encodedPayload is scale encoded of payload in hex format",
            "example": "0x012345"
          }
        },
        "required": [
          "payload",
          "encodedPayload"
        ]
      },
      "KeysRequestPayloadDto": {
        "type": "object",
        "properties": {
          "msaId": {
            "type": "string",
            "description": "MSA Id of the user requesting the new key",
            "example": "3"
          },
          "expiration": {
            "type": "number",
            "description": "expiration block number for this payload",
            "example": 1
          },
          "newPublicKey": {
            "type": "string",
            "description": "newPublicKey in hex format",
            "example": "0x0ed2f8c714efcac51ca2325cfe95637e5e0b898ae397aa365978b7348a717d0b"
          }
        },
        "required": [
          "msaId",
          "expiration",
          "newPublicKey"
        ]
      },
      "KeysRequestDto": {
        "type": "object",
        "properties": {
          "msaOwnerAddress": {
            "type": "string",
            "description": "msaOwnerAddress representing the target of this request",
            "example": "1LSLqpLWXo7A7xuiRdu6AQPnBPNJHoQSu8DBsUYJgsNEJ4N"
          },
          "msaOwnerSignature": {
            "type": "string",
            "description": "msaOwnerSignature is the signature by msa owner",
            "example": "0x065d733ca151c9e65b78f2ba77348224d31647e6913c44ad2765c6e8ba06f834dc21d8182447d01c30f84a41d90a8f2e58001d825c6f0d61b0afe89f984eec85"
          },
          "newKeyOwnerSignature": {
            "type": "string",
            "description": "newKeyOwnerSignature is the signature with new key",
            "example": "0x065d733ca151c9e65b78f2ba77348224d31647e6913c44ad2765c6e8ba06f834dc21d8182447d01c30f84a41d90a8f2e58001d825c6f0d61b0afe89f984eec85"
          },
          "payload": {
            "$ref": "#/components/schemas/KeysRequestPayloadDto"
          }
        },
        "required": [
          "msaOwnerAddress",
          "msaOwnerSignature",
          "newKeyOwnerSignature",
          "payload"
        ]
      },
      "KeysResponse": {
        "type": "object",
        "properties": {
          "msaKeys": {
            "type": "object"
          }
        },
        "required": [
          "msaKeys"
        ]
      },
      "ItemActionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Action Item type",
            "example": "ADD_ITEM",
            "enum": [
              "ADD_ITEM",
              "DELETE_ITEM"
            ]
          },
          "encodedPayload": {
            "type": "string",
            "description": "encodedPayload to be added",
            "example": "0x1234"
          },
          "index": {
            "type": "number",
            "description": "index of the item to be deleted",
            "example": 0
          }
        },
        "required": [
          "type"
        ]
      },
      "ItemizedSignaturePayloadDto": {
        "type": "object",
        "properties": {
          "actions": {
            "example": [
              {
                "type": "ADD_ITEM",
                "encodedPayload": "0x1122"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemActionDto"
            }
          },
          "schemaId": {
            "type": "number",
            "description": "schemaId related to the payload",
            "example": 1
          },
          "targetHash": {
            "type": "number",
            "description": "targetHash related to the stateful storage",
            "example": 1234
          },
          "expiration": {
            "type": "number",
            "description": "expiration block number for this payload",
            "example": 1
          }
        },
        "required": [
          "actions",
          "schemaId",
          "targetHash",
          "expiration"
        ]
      },
      "AddNewPublicKeyAgreementPayloadRequest": {
        "type": "object",
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/ItemizedSignaturePayloadDto"
          },
          "encodedPayload": {
            "type": "string",
            "description": "Raw encodedPayload to be signed",
            "example": "0x1234"
          }
        },
        "required": [
          "payload",
          "encodedPayload"
        ]
      },
      "AddNewPublicKeyAgreementRequestDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "description": "AccountId in hex or SS58 format",
            "example": "1LSLqpLWXo7A7xuiRdu6AQPnBPNJHoQSu8DBsUYJgsNEJ4N"
          },
          "payload": {
            "$ref": "#/components/schemas/ItemizedSignaturePayloadDto"
          },
          "proof": {
            "type": "string",
            "description": "proof is the signature for the payload",
            "example": "0x065d733ca151c9e65b78f2ba77348224d31647e6913c44ad2765c6e8ba06f834dc21d8182447d01c30f84a41d90a8f2e58001d825c6f0d61b0afe89f984eec85"
          }
        },
        "required": [
          "accountId",
          "payload",
          "proof"
        ]
      }
    }
  }
}